[tlv] Split out classes TLVReader, TLVWriter, TLVUpdater, and TLVBackingStore.
